Fixed interface directory recognition

+++ config.php
@@ -41,13 +41,17 @@ $instance_configuration_path = $config["instance_directory"] . "/config.json"; /
 if (file_exists($instance_configuration_path)) { // Check to see if the instance configuration file exists.
     $instance_config = json_decode(file_get_contents($config["instance_directory"] . "/config.json"), true); // Load the instance configuration file.
     if ($instance_config["external"]["local"]["enabled"] == true) {
-        $_POST["interface_directory"] = $instance_config["external"]["local"]["interface_directory"]; // Auto-fill the interface directory.
+        $config["interface_directory"] = $instance_config["external"]["local"]["interface_directory"]; // Auto-fill the interface directory.
     } else {
         echo "<p class=\"error\">The local interface directory is disabled in Assassin's configuration. " . htmlspecialchars($config["product_name"]) . " requires this feature to be enabled to function.</p>"; // Inform the user that the database failed to load.
         exit(); // Terminate the script.
     }
 } else {
-    echo "<p class=\"error\">The interface directory could not be identified from the instance configuration file. It is possible the Assassin configuration file is corrupt, or the incorrect Assassin instance directory is set.</p>"; // Inform the user that the database failed to load.
+    if (!isset($displayed_instance_config_warning)) { // Only show the instance configuration warning if it has not already been displayed.
+        echo "<p class=\"error\">The interface directory could not be identified from the instance configuration file. It is possible the Assassin configuration file is corrupt, or the incorrect Assassin instance directory is set.</p>"; // Inform the user that the database failed to load.
+    }
+    $displayed_instance_config_warning = true; // Indicate that the instance configuration warning has already been displayed.
+
 }
 
 ?>
